+++ tests/BaseTestCase.php
@@ -53,6 +53,7 @@ abstract class BaseTestCase extends OrchestraTestCase
                 'event' => 'Events',
                 'listener' => 'Listeners',
                 'migration' => 'Database/Migrations',
+                'factory' => 'Database/factories',
                 'model' => 'Entities',
                 'repository' => 'Repositories',
                 'seeder' => 'Database/Seeders',